+++ dev/tests/functional/tests/app/Magento/SalesRule/Test/TestStep/CreateSalesRuleStep.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright © 2015 Magento. All rights reserved.
+ * Copyright © 2016 Magento. All rights reserved.
  * See COPYING.txt for license details.
  */
 
@@ -59,9 +59,10 @@ class CreateSalesRuleStep implements TestStepInterface
     {
         $result['salesRule'] = null;
         if ($this->salesRule !== null) {
+            $this->deleteAllSalesRule->run();
             $salesRule = $this->fixtureFactory->createByCode(
                 'salesRule',
-                ['dataSet' => $this->salesRule]
+                ['dataset' => $this->salesRule]
             );
             $salesRule->persist();
             $result['salesRule'] = $salesRule;
@@ -77,6 +78,8 @@ class CreateSalesRuleStep implements TestStepInterface
      */
     public function cleanup()
     {
-        $this->deleteAllSalesRule->run();
+        if ($this->salesRule !== null) {
+            $this->deleteAllSalesRule->run();
+        }
     }
 }
