+++ dev/tests/functional/tests/app/Magento/Tax/Test/TestStep/CreateTaxRuleStep.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright © 2015 Magento. All rights reserved.
+ * Copyright © 2016 Magento. All rights reserved.
  * See COPYING.txt for license details.
  */
 
@@ -59,11 +59,12 @@ class CreateTaxRuleStep implements TestStepInterface
     {
         $result['taxRule'] = null;
         if ($this->taxRule !== null) {
+            $this->deleteAllTaxRule->run();
             $taxRuleDataSets = explode(',', $this->taxRule);
             foreach ($taxRuleDataSets as $taxRuleDataSet) {
                 $taxRule = $this->fixtureFactory->createByCode(
                     'taxRule',
-                    ['dataSet' => $taxRuleDataSet]
+                    ['dataset' => $taxRuleDataSet]
                 );
                 $taxRule->persist();
                 $result['taxRule'] = $taxRule;
@@ -80,6 +81,8 @@ class CreateTaxRuleStep implements TestStepInterface
      */
     public function cleanup()
     {
-        $this->deleteAllTaxRule->run();
+        if ($this->taxRule !== null) {
+            $this->deleteAllTaxRule->run();
+        }
     }
 }
